+++ client/state/selectors/is-google-my-business-stats-nudge-visible.js
@@ -1,59 +1,23 @@
 /** @format */
 
 /**
- * External dependencies
- */
-import config from 'config';
-
-/**
  * Internal dependencies
  */
-import createSelector from 'lib/create-selector';
-import { getSiteOption, getSitePlanSlug } from 'state/sites/selectors';
-import { isGoogleMyBusinessLocationConnected } from 'state/selectors';
-import { isRequestingSiteSettings, getSiteSettings } from 'state/site-settings/selectors';
-import { planMatches } from 'lib/plans';
-import { TYPE_BUSINESS, GROUP_WPCOM } from 'lib/plans/constants';
-
-/**
- * Returns true if site has promote goal set
- *
- * @param  {Object}  state  Global state tree
- * @param  {String}  siteId The Site ID
- * @return {Boolean} True if site has "promote" goal
- */
-const siteHasPromoteGoal = createSelector(
-	( state, siteId ) => {
-		const siteGoals = ( getSiteOption( state, siteId, 'site_goals' ) || '' ).split( ',' );
-
-		return siteGoals.indexOf( 'promote' ) !== -1;
-	},
-	( state, siteId ) => [ getSiteOption( state, siteId, 'site_goals' ) ]
-);
-
-/**
- * Returns true if site has business plan
- *
- * @param  {Object}  state  Global state tree
- * @param  {String}  siteId The Site ID
- * @return {Boolean} True if site has business plan
- */
-export const siteHasBusinessPlan = createSelector(
-	( state, siteId ) => {
-		const slug = getSitePlanSlug( state, siteId );
-
-		return planMatches( slug, { group: GROUP_WPCOM, type: TYPE_BUSINESS } );
-	},
-	( state, siteId ) => [ getSitePlanSlug( state, siteId ) ]
-);
+import {
+	isGoogleMyBusinessLocationConnected,
+	isSiteGoogleMyBusinessEligible,
+} from 'state/selectors';
+import {
+	isRequestingSiteSettings,
+	getSiteSettings,
+} from 'state/site-settings/selectors';
 
 /**
  * Returns true if the Google My Business (GMB) nudge should be visible in stats
  *
  * It should be visible if:
- * - site is older than 1 week,
- * - site has a business plan
- * - site has a promote goal
+ * - it meets the Google My Business Site Eligiblility Critera ( see isSiteGoogleMyBusinessEligible ),
+ * - site has NOT been connected to a location
  * @param  {Object}  state  Global state tree
  * @param  {String}  siteId The Site ID
  * @return {Boolean} True if we should show the nudge
@@ -70,10 +34,5 @@ export default function isGoogleMyBusinessStatsNudgeVisible( state, siteId ) {
 		return false;
 	}
 
-	// call-for-testing condition, remove on launch
-	if ( config.isEnabled( 'google-my-business' ) ) {
-		return siteHasBusinessPlan( state, siteId );
-	}
-
-	return siteHasBusinessPlan( state, siteId ) && siteHasPromoteGoal( state, siteId );
+	return isSiteGoogleMyBusinessEligible( state, siteId );
 }
